+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+## [0.1.0] - 2018-02-15
+### Added
+    - Add errors after creating product without name
+    - Test for product validation
+    - Chanre respons of most part requests for errors chechs
+### Fixed
+    - Fix errors after name change
 ## [0.0.1] - 2018-01-30
 ### Fixed
     - Fixed error after uncorect product name from edit product method